backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
motor==3.4.0
pymongo==4.6.3
pydantic==2.12.5
email-validator==2.3.0
PyJWT==2.10.1
bcrypt==4.1.3
python-multipart==0.0.20
scikit-learn==1.5.2
joblib==1.4.2
httpx==0.28.1
transformers==4.52.4
Pillow==11.2.1
torch==2.7.1
